restoran/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
